Add in KMP UUID library for Session Vault Salt
