+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/AuthChallenge.kt
@@ -1,0 +1,7 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class AuthChallenge(
+    val challenge: String,
+    val sessionId: String,
+    val expiresAt: Long,
+)
