+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/BudgetCap.kt
@@ -1,0 +1,6 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class BudgetCap(
+    val amount: String,
+    val asset: String,
+)
