+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/ClientConfig.kt
@@ -1,0 +1,7 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class ClientConfig(
+    val autoPaySegments: Boolean = false,
+    val maxAutoPaySegments: Int? = null,
+    val budgetCap: BudgetCap? = null,
+)
