+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/ConsentTerms.kt
@@ -1,0 +1,12 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class ConsentTerms(
+    val gatingMode: GatingMode,
+    val amount: String,
+    val asset: String,
+    val network: String,
+    val payTo: String? = null,
+    val segmentDuration: Int? = null,
+    val segmentBytes: Long? = null,
+    val suggestedBudgetCap: String? = null,
+)
