+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/EnforcementMode.kt
@@ -1,0 +1,8 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+enum class EnforcementMode(
+    val value: String,
+) {
+    TRACK("track"),
+    CRYPTO("crypto"),
+}
