+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/PaymentReceipt.kt
@@ -1,0 +1,16 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class PaymentReceipt(
+    val txId: String,
+    val sessionId: String,
+    val segmentIndex: Int,
+    val amount: String,
+    val asset: String,
+    val payTo: String,
+    val payFrom: String = "",
+    val feePayer: String? = null,
+    val facilitator: String? = null,
+    val network: String,
+    val timestamp: Long,
+    val channelId: String? = null,
+)
