+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/PaymentRequestMeta.kt
@@ -1,0 +1,10 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class PaymentRequestMeta(
+    val gatingMode: GatingMode,
+    val enforcement: EnforcementMode,
+    val segmentDuration: Int? = null,
+    val segmentBytes: Long? = null,
+    val viewerAddress: String? = null,
+    val voucherSignature: String? = null,
+)
