+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/RailPayment.kt
@@ -1,0 +1,9 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class RailPayment(
+    val railId: String,
+    val version: Int,
+    val nonce: String,
+    val paymentPayload: Any,
+    val paymentRequirements: Any,
+)
