+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/SessionStats.kt
@@ -1,0 +1,10 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class SessionStats(
+    val sessionId: String,
+    var segmentsDelivered: Int = 0,
+    var segmentsPaid: Int = 0,
+    var totalAmountReceived: String = "0",
+    var totalBytesTransferred: Long = 0,
+    var durationSeconds: Long = 0,
+)
