+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/domain/model/SpendTransaction.kt
@@ -1,0 +1,8 @@
+package com.michaeltchuang.walletsdk.core.railmpp.domain.model
+
+data class SpendTransaction(
+    val txId: String,
+    val amount: String,
+    val segmentIndex: Int,
+    val timestamp: Long,
+)
