+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/internal/AlgorandOps.kt
@@ -17,7 +17,6 @@ internal expect suspend fun submitAppCallInternal(
     signer: MppWalletSigner,
     appId: Long,
     usdcAssetId: Long,
-    defaultSalt: ByteArray,
     algodUrl: String,
     args: List<ByteArray>,
     boxKeys: List<Pair<Long, ByteArray>>,
