+++ wallet-sdk-core/src/commonMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/smartcontract/EscrowSessionVaultManagerClient.kt
@@ -16,7 +16,7 @@ import com.michaeltchuang.walletsdk.core.railmpp.utils.RailMppConstants
 class EscrowSessionVaultManagerClient(
     private val appId: Long = RailMppConstants.MPP_SESSION_VAULT_APP_ID,
     private val usdcAssetId: Long = AssetConstants.USDC_TESTNET_ID,
-    private val defaultSalt: ByteArray,
+    val defaultSalt: ByteArray,
     private val defaultAlgodUrl: String = DEFAULT_ALGOD_URL,
 ) {
     companion object {
@@ -37,6 +37,7 @@ class EscrowSessionVaultManagerClient(
         private val ABI_OPT_IN_USDC = byteArrayOf(0x7e, 0x3f, 0x4a, 0x68)
         private val ABI_VERIFY_SETTLE_SIGNATURE = byteArrayOf(0x27, 0x04, 0x92.toByte(), 0x89.toByte())
         var channelId: ByteArray? = null
+        var salt: ByteArray? = null
     }
 
     fun deriveChannelId(
@@ -74,6 +75,7 @@ class EscrowSessionVaultManagerClient(
                 "payerAddress must match signer.address for session vault deposit"
             }
             val channelId = channelId ?: error("channelId is null")
+            val salt = salt ?: error("salt is null")
 
             submitAssetTransferAndAppCallInternal(
                 signer = signer,
@@ -84,7 +86,7 @@ class EscrowSessionVaultManagerClient(
                     listOf(
                         ABI_OPEN,
                         decodeAlgorandAddressPublicKey(payeeAddress),
-                        encodeArc4DynamicBytes(defaultSalt),
+                        encodeArc4DynamicBytes(salt),
                         encodeArc4DynamicBytes(computeSignerPubkeyHash(signer.authorizedSignerPublicKey)),
                         encodeArc4DynamicBytes(signer.authorizedSignerPublicKey),
                     ),
@@ -128,7 +130,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args =
                     listOf(
@@ -157,7 +158,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args =
                     listOf(
@@ -187,7 +187,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args =
                     listOf(
@@ -215,7 +214,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_SETTLE_LATEST, encodeArc4DynamicBytes(channelId)),
                 boxKeys = listOf(Pair(appId, channelId)),
@@ -234,7 +232,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_CLOSE, encodeArc4DynamicBytes(channelId)),
                 boxKeys = listOf(Pair(appId, channelId)),
@@ -254,7 +251,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_REQUEST_CLOSE, encodeArc4DynamicBytes(channelId)),
                 boxKeys = listOf(Pair(appId, channelId)),
@@ -272,7 +268,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_WITHDRAW, encodeArc4DynamicBytes(channelId)),
                 boxKeys = listOf(Pair(appId, channelId)),
@@ -291,7 +286,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_FUND_MBR_POOL, decodeAlgorandAddressPublicKey(receiverAddress)),
                 boxKeys = emptyList(),
@@ -308,7 +302,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args = listOf(ABI_OPT_IN_USDC),
                 boxKeys = emptyList(),
@@ -328,7 +321,6 @@ class EscrowSessionVaultManagerClient(
                 signer = signer,
                 appId = appId,
                 usdcAssetId = usdcAssetId,
-                defaultSalt = defaultSalt,
                 algodUrl = algodUrl,
                 args =
                     listOf(
