+++ wallet-sdk-core/src/iosMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/internal/AlgorandOps.ios.kt
@@ -47,7 +47,6 @@ internal actual suspend fun submitAppCallInternal(
     signer: MppWalletSigner,
     appId: Long,
     usdcAssetId: Long,
-    defaultSalt: ByteArray,
     algodUrl: String,
     args: List<ByteArray>,
     boxKeys: List<Pair<Long, ByteArray>>,
